+++ src/frameworks/next/index.ts
@@ -1,11 +1,16 @@
 import { execSync } from "child_process";
-import { readFile, mkdir, copyFile } from "fs/promises";
+import { mkdir, copyFile } from "fs/promises";
 import { dirname, join } from "path";
-import type { Header, Rewrite, Redirect } from "next/dist/lib/load-custom-routes";
 import type { NextConfig } from "next";
+import type { PrerenderManifest } from "next/dist/build";
+import type { MiddlewareManifest } from "next/dist/build/webpack/plugins/middleware-plugin";
+import type { PagesManifest } from "next/dist/build/webpack/plugins/pages-manifest-plugin";
 import { copy, mkdirp, pathExists } from "fs-extra";
 import { pathToFileURL, parse } from "url";
 import { existsSync } from "fs";
+import { gte } from "semver";
+import { IncomingMessage, ServerResponse } from "http";
+import * as clc from "colorette";
 
 import {
   BuildResult,
@@ -17,37 +22,38 @@ import {
   SupportLevel,
 } from "..";
 import { promptOnce } from "../../prompt";
-import { gte } from "semver";
-import { IncomingMessage, ServerResponse } from "http";
 import { logger } from "../../logger";
 import { FirebaseError } from "../../error";
-import { fileExistsSync } from "../../fsutils";
+import {
+  cleanEscapedChars,
+  getNextjsRewritesToUse,
+  isHeaderSupportedByHosting,
+  isRedirectSupportedByHosting,
+  isRewriteSupportedByHosting,
+  isUsingAppDirectory,
+  isUsingImageOptimization,
+  isUsingMiddleware,
+} from "./utils";
+import type { Manifest } from "./interfaces";
+import { readJSON } from "../utils";
+import { warnIfCustomBuildScript } from "../utils";
+import type { EmulatorInfo } from "../../emulator/types";
+import { usesAppDirRouter, usesNextImage, hasUnoptimizedImage } from "./utils";
+import {
+  MIDDLEWARE_MANIFEST,
+  PAGES_MANIFEST,
+  PRERENDER_MANIFEST,
+  ROUTES_MANIFEST,
+} from "./constants";
 
-// Next.js's exposed interface is incomplete here
-// TODO see if there's a better way to grab this
-interface Manifest {
-  distDir?: string;
-  basePath?: string;
-  headers?: (Header & { regex: string })[];
-  redirects?: (Redirect & { regex: string })[];
-  rewrites?:
-    | (Rewrite & { regex: string })[]
-    | {
-        beforeFiles?: (Rewrite & { regex: string })[];
-        afterFiles?: (Rewrite & { regex: string })[];
-        fallback?: (Rewrite & { regex: string })[];
-      };
-}
-
-const CLI_COMMAND = join(
-  "node_modules",
-  ".bin",
-  process.platform === "win32" ? "next.cmd" : "next"
-);
+const DEFAULT_BUILD_SCRIPT = ["next build"];
+const PUBLIC_DIR = "public";
 
 export const name = "Next.js";
 export const support = SupportLevel.Experimental;
 export const type = FrameworkType.MetaFramework;
+
+const DEFAULT_NUMBER_OF_REASONS_TO_LIST = 5;
 
 function getNextVersion(cwd: string): string | undefined {
   return findDependency("next", { cwd, depth: 0, omitDev: false })?.version;
@@ -63,8 +69,8 @@ function getReactVersion(cwd: string): string | undefined {
 export async function discover(dir: string) {
   if (!(await pathExists(join(dir, "package.json")))) return;
   if (!(await pathExists("next.config.js")) && !getNextVersion(dir)) return;
-  // TODO don't hardcode public dir
-  return { mayWantBackend: true, publicDirectory: join(dir, "public") };
+
+  return { mayWantBackend: true, publicDirectory: join(dir, PUBLIC_DIR) };
 }
 
 /**
@@ -72,6 +78,8 @@ export async function discover(dir: string) {
  */
 export async function build(dir: string): Promise<BuildResult> {
   const { default: nextBuild } = relativeRequire(dir, "next/dist/build");
+
+  await warnIfCustomBuildScript(dir, name, DEFAULT_BUILD_SCRIPT);
 
   const reactVersion = getReactVersion(dir);
   if (reactVersion && gte(reactVersion, "18.0.0")) {
@@ -86,76 +94,138 @@ export async function build(dir: string): Promise<BuildResult> {
     throw e;
   });
 
-  try {
-    // Using spawn here, rather than their programatic API because I can't silence it
-    // Failures with Next export are expected, we're just trying to do it if we can
-    execSync(`${CLI_COMMAND} export`, { cwd: dir, stdio: "ignore" });
-  } catch (e) {
-    // continue, failure is expected
+  const reasonsForBackend = [];
+  const { distDir } = await getConfig(dir);
+
+  if (await isUsingMiddleware(join(dir, distDir), false)) {
+    reasonsForBackend.push("middleware");
   }
 
-  let wantsBackend = true;
-  const { distDir } = await getConfig(dir);
-  const exportDetailPath = join(dir, distDir, "export-detail.json");
-  const exportDetailExists = await pathExists(exportDetailPath);
-  const exportDetailBuffer = exportDetailExists ? await readFile(exportDetailPath) : undefined;
-  const exportDetailJson = exportDetailBuffer && JSON.parse(exportDetailBuffer.toString());
-  if (exportDetailJson?.success) {
-    const appPathRoutesManifestPath = join(dir, distDir, "app-path-routes-manifest.json");
-    const appPathRoutesManifestJSON = fileExistsSync(appPathRoutesManifestPath)
-      ? await readFile(appPathRoutesManifestPath).then((it) => JSON.parse(it.toString()))
-      : {};
-    const prerenderManifestJSON = await readFile(
-      join(dir, distDir, "prerender-manifest.json")
-    ).then((it) => JSON.parse(it.toString()));
-    const anyDynamicRouteFallbacks = !!Object.values(
-      prerenderManifestJSON.dynamicRoutes || {}
-    ).find((it: any) => it.fallback !== false);
-    const pagesManifestJSON = await readFile(
-      join(dir, distDir, "server", "pages-manifest.json")
-    ).then((it) => JSON.parse(it.toString()));
-    const prerenderedRoutes = Object.keys(prerenderManifestJSON.routes);
-    const dynamicRoutes = Object.keys(prerenderManifestJSON.dynamicRoutes);
-    const unrenderedPages = [
-      ...Object.keys(pagesManifestJSON),
-      // TODO flush out fully rendered detection with a app directory (Next 13)
-      // we shouldn't go too crazy here yet, as this is currently an expiriment
-      ...Object.values<string>(appPathRoutesManifestJSON),
-    ].filter(
-      (it) =>
-        !(
-          ["/_app", "/", "/_error", "/_document", "/404"].includes(it) ||
-          prerenderedRoutes.includes(it) ||
-          dynamicRoutes.includes(it)
-        )
-    );
-    // TODO log these as a reason why Cloud Functions are needed
-    if (!anyDynamicRouteFallbacks && unrenderedPages.length === 0) {
-      wantsBackend = false;
+  if (await isUsingImageOptimization(join(dir, distDir))) {
+    reasonsForBackend.push(`Image Optimization`);
+  }
+
+  if (isUsingAppDirectory(join(dir, distDir))) {
+    // Let's not get smart here, if they are using the app directory we should
+    // opt for spinning up a Cloud Function. The app directory is unstable.
+    reasonsForBackend.push("app directory (unstable)");
+  }
+
+  const prerenderManifest = await readJSON<PrerenderManifest>(
+    join(dir, distDir, PRERENDER_MANIFEST)
+  );
+
+  const dynamicRoutesWithFallback = Object.entries(prerenderManifest.dynamicRoutes || {}).filter(
+    ([, it]) => it.fallback !== false
+  );
+  if (dynamicRoutesWithFallback.length > 0) {
+    for (const [key] of dynamicRoutesWithFallback) {
+      reasonsForBackend.push(`use of fallback ${key}`);
     }
   }
 
-  const manifestBuffer = await readFile(join(dir, distDir, "routes-manifest.json"));
-  const manifest: Manifest = JSON.parse(manifestBuffer.toString());
+  const routesWithRevalidate = Object.entries(prerenderManifest.routes).filter(
+    ([, it]) => it.initialRevalidateSeconds
+  );
+  if (routesWithRevalidate.length > 0) {
+    for (const [key] of routesWithRevalidate) {
+      reasonsForBackend.push(`use of revalidate ${key}`);
+    }
+  }
+
+  const pagesManifestJSON = await readJSON<PagesManifest>(
+    join(dir, distDir, "server", PAGES_MANIFEST)
+  );
+  const prerenderedRoutes = Object.keys(prerenderManifest.routes);
+  const dynamicRoutes = Object.keys(prerenderManifest.dynamicRoutes);
+  const unrenderedPages = Object.keys(pagesManifestJSON).filter(
+    (it) =>
+      !(
+        ["/_app", "/", "/_error", "/_document", "/404"].includes(it) ||
+        prerenderedRoutes.includes(it) ||
+        dynamicRoutes.includes(it)
+      )
+  );
+  if (unrenderedPages.length > 0) {
+    for (const key of unrenderedPages) {
+      reasonsForBackend.push(`non-static route ${key}`);
+    }
+  }
+
+  const manifest = await readJSON<Manifest>(join(dir, distDir, ROUTES_MANIFEST));
+
   const {
     headers: nextJsHeaders = [],
     redirects: nextJsRedirects = [],
     rewrites: nextJsRewrites = [],
   } = manifest;
-  const headers = nextJsHeaders.map(({ source, headers }) => ({ source, headers }));
+
+  const isEveryHeaderSupported = nextJsHeaders.every(isHeaderSupportedByHosting);
+  if (!isEveryHeaderSupported) {
+    reasonsForBackend.push("advanced headers");
+  }
+
+  const headers = nextJsHeaders.filter(isHeaderSupportedByHosting).map(({ source, headers }) => ({
+    // clean up unnecessary escaping
+    source: cleanEscapedChars(source),
+    headers,
+  }));
+
+  const isEveryRedirectSupported = nextJsRedirects.every(isRedirectSupportedByHosting);
+  if (!isEveryRedirectSupported) {
+    reasonsForBackend.push("advanced redirects");
+  }
+
   const redirects = nextJsRedirects
-    .filter(({ internal }: any) => !internal)
-    .map(({ source, destination, statusCode: type }) => ({ source, destination, type }));
-  const nextJsRewritesToUse = Array.isArray(nextJsRewrites)
-    ? nextJsRewrites
-    : nextJsRewrites.beforeFiles || [];
+    .filter(isRedirectSupportedByHosting)
+    .map(({ source, destination, statusCode: type }) => ({
+      // clean up unnecessary escaping
+      source: cleanEscapedChars(source),
+      destination,
+      type,
+    }));
+
+  const nextJsRewritesToUse = getNextjsRewritesToUse(nextJsRewrites);
+
+  // rewrites.afterFiles / rewrites.fallback are not supported by firebase.json
+  if (
+    !Array.isArray(nextJsRewrites) &&
+    (nextJsRewrites.afterFiles?.length || nextJsRewrites.fallback?.length)
+  ) {
+    reasonsForBackend.push("advanced rewrites");
+  }
+
+  const isEveryRewriteSupported = nextJsRewritesToUse.every(isRewriteSupportedByHosting);
+  if (!isEveryRewriteSupported) {
+    reasonsForBackend.push("advanced rewrites");
+  }
+
+  // Can we change i18n into Firebase settings?
   const rewrites = nextJsRewritesToUse
-    .map(({ source, destination, has }) => {
-      // Can we change i18n into Firebase settings?
-      if (has) return undefined;
-      return { source, destination };
-    })
-    .filter((it) => it);
+    .filter(isRewriteSupportedByHosting)
+    .map(({ source, destination }) => ({
+      // clean up unnecessary escaping
+      source: cleanEscapedChars(source),
+      destination,
+    }));
+
+  const wantsBackend = reasonsForBackend.length > 0;
+
+  if (wantsBackend) {
+    const numberOfReasonsToList = process.env.DEBUG ? Infinity : DEFAULT_NUMBER_OF_REASONS_TO_LIST;
+    console.log("Building a Cloud Function to run this application. This is needed due to:");
+    for (const reason of reasonsForBackend.slice(0, numberOfReasonsToList)) {
+      console.log(` • ${reason}`);
+    }
+    if (reasonsForBackend.length > numberOfReasonsToList) {
+      console.log(
+        ` • and ${
+          reasonsForBackend.length - numberOfReasonsToList
+        } other reasons, use --debug to see more`
+      );
+    }
+    console.log("");
+  }
 
   return { wantsBackend, headers, redirects, rewrites };
 }
@@ -183,79 +253,86 @@ export async function init(setup: any) {
  */
 export async function ɵcodegenPublicDirectory(sourceDir: string, destDir: string) {
   const { distDir } = await getConfig(sourceDir);
-  const exportDetailPath = join(sourceDir, distDir, "export-detail.json");
-  const exportDetailExists = await pathExists(exportDetailPath);
-  const exportDetailBuffer = exportDetailExists ? await readFile(exportDetailPath) : undefined;
-  const exportDetailJson = exportDetailBuffer && JSON.parse(exportDetailBuffer.toString());
-  if (exportDetailJson?.success) {
-    copy(exportDetailJson.outDirectory, destDir);
-  } else {
-    const publicPath = join(sourceDir, "public");
-    await mkdir(join(destDir, "_next", "static"), { recursive: true });
-    if (await pathExists(publicPath)) {
-      await copy(publicPath, destDir);
-    }
-    await copy(join(sourceDir, distDir, "static"), join(destDir, "_next", "static"));
 
-    // Copy over the default html files
-    for (const file of ["index.html", "404.html", "500.html"]) {
-      const pagesPath = join(sourceDir, distDir, "server", "pages", file);
-      if (await pathExists(pagesPath)) {
-        await copyFile(pagesPath, join(destDir, file));
-        continue;
-      }
-      const appPath = join(sourceDir, distDir, "server", "app", file);
-      if (await pathExists(appPath)) {
-        await copyFile(appPath, join(destDir, file));
-      }
+  const publicPath = join(sourceDir, "public");
+  await mkdir(join(destDir, "_next", "static"), { recursive: true });
+  if (await pathExists(publicPath)) {
+    await copy(publicPath, destDir);
+  }
+  await copy(join(sourceDir, distDir, "static"), join(destDir, "_next", "static"));
+
+  // Copy over the default html files
+  for (const file of ["index.html", "404.html", "500.html"]) {
+    const pagesPath = join(sourceDir, distDir, "server", "pages", file);
+    if (await pathExists(pagesPath)) {
+      await copyFile(pagesPath, join(destDir, file));
+      continue;
+    }
+    const appPath = join(sourceDir, distDir, "server", "app", file);
+    if (await pathExists(appPath)) {
+      await copyFile(appPath, join(destDir, file));
+    }
+  }
+
+  const [middlewareManifest, prerenderManifest, routesManifest] = await Promise.all([
+    readJSON<MiddlewareManifest>(join(sourceDir, distDir, "server", MIDDLEWARE_MANIFEST)),
+    readJSON<PrerenderManifest>(join(sourceDir, distDir, PRERENDER_MANIFEST)),
+    readJSON<Manifest>(join(sourceDir, distDir, ROUTES_MANIFEST)),
+  ]);
+
+  const middlewareMatcherRegexes = Object.values(middlewareManifest.middleware)
+    .map((it) => it.matchers)
+    .flat()
+    .map((it) => new RegExp(it.regexp));
+
+  const { redirects = [], rewrites = [], headers = [] } = routesManifest;
+
+  const rewritesRegexesNotSupportedByHosting = getNextjsRewritesToUse(rewrites)
+    .filter((rewrite) => !isRewriteSupportedByHosting(rewrite))
+    .map((rewrite) => new RegExp(rewrite.regex));
+
+  const redirectsRegexesNotSupportedByHosting = redirects
+    .filter((redirect) => !isRedirectSupportedByHosting(redirect))
+    .map((redirect) => new RegExp(redirect.regex));
+
+  const headersRegexesNotSupportedByHosting = headers
+    .filter((header) => !isHeaderSupportedByHosting(header))
+    .map((header) => new RegExp(header.regex));
+
+  const pathsUsingsFeaturesNotSupportedByHosting = [
+    ...middlewareMatcherRegexes,
+    ...rewritesRegexesNotSupportedByHosting,
+    ...redirectsRegexesNotSupportedByHosting,
+    ...headersRegexesNotSupportedByHosting,
+  ];
+
+  for (const [path, route] of Object.entries(prerenderManifest.routes)) {
+    if (
+      route.initialRevalidateSeconds ||
+      pathsUsingsFeaturesNotSupportedByHosting.some((it) => path.match(it))
+    ) {
+      continue;
     }
 
-    const prerenderManifestBuffer = await readFile(
-      join(sourceDir, distDir, "prerender-manifest.json")
+    const isReactServerComponent = route.dataRoute.endsWith(".rsc");
+    const contentDist = join(
+      sourceDir,
+      distDir,
+      "server",
+      isReactServerComponent ? "app" : "pages"
     );
-    const prerenderManifest = JSON.parse(prerenderManifestBuffer.toString());
-    for (const path in prerenderManifest.routes) {
-      if (prerenderManifest.routes[path]) {
-        // Skip ISR in the deploy to hosting
-        const { initialRevalidateSeconds } = prerenderManifest.routes[path];
-        if (initialRevalidateSeconds) {
-          continue;
-        }
 
-        // TODO(jamesdaniels) explore oppertunity to simplify this now that we
-        //                    are defaulting cleanURLs to true for frameworks
+    const parts = path.split("/").filter((it) => !!it);
+    const partsOrIndex = parts.length > 0 ? parts : ["index"];
 
-        // / => index.json => index.html => index.html
-        // /foo => foo.json => foo.html
-        const parts = path
-          .split("/")
-          .slice(1)
-          .filter((it) => !!it);
-        const partsOrIndex = parts.length > 0 ? parts : ["index"];
-        const dataPath = `${join(...partsOrIndex)}.json`;
-        const htmlPath = `${join(...partsOrIndex)}.html`;
-        await mkdir(join(destDir, dirname(htmlPath)), { recursive: true });
-        const pagesHtmlPath = join(sourceDir, distDir, "server", "pages", htmlPath);
-        if (await pathExists(pagesHtmlPath)) {
-          await copyFile(pagesHtmlPath, join(destDir, htmlPath));
-        } else {
-          const appHtmlPath = join(sourceDir, distDir, "server", "app", htmlPath);
-          if (await pathExists(appHtmlPath)) {
-            await copyFile(appHtmlPath, join(destDir, htmlPath));
-          }
-        }
-        const dataRoute = prerenderManifest.routes[path].dataRoute;
-        await mkdir(join(destDir, dirname(dataRoute)), { recursive: true });
-        const pagesDataPath = join(sourceDir, distDir, "server", "pages", dataPath);
-        if (await pathExists(pagesDataPath)) {
-          await copyFile(pagesDataPath, join(destDir, dataRoute));
-        } else {
-          const appDataPath = join(sourceDir, distDir, "server", "app", dataPath);
-          if (await pathExists(appDataPath)) {
-            await copyFile(appDataPath, join(destDir, dataRoute));
-          }
-        }
-      }
+    const htmlPath = `${join(...partsOrIndex)}.html`;
+    await mkdir(join(destDir, dirname(htmlPath)), { recursive: true });
+    await copyFile(join(contentDist, htmlPath), join(destDir, htmlPath));
+
+    if (!isReactServerComponent) {
+      const dataPath = `${join(...partsOrIndex)}.json`;
+      await mkdir(join(destDir, dirname(route.dataRoute)), { recursive: true });
+      await copyFile(join(contentDist, dataPath), join(destDir, route.dataRoute));
     }
   }
 }
@@ -265,8 +342,7 @@ export async function ɵcodegenPublicDirectory(sourceDir: string, destDir: strin
  */
 export async function ɵcodegenFunctionsDirectory(sourceDir: string, destDir: string) {
   const { distDir } = await getConfig(sourceDir);
-  const packageJsonBuffer = await readFile(join(sourceDir, "package.json"));
-  const packageJson = JSON.parse(packageJsonBuffer.toString());
+  const packageJson = await readJSON(join(sourceDir, "package.json"));
   if (existsSync(join(sourceDir, "next.config.js"))) {
     let esbuild;
     try {
@@ -292,6 +368,18 @@ export async function ɵcodegenFunctionsDirectory(sourceDir: string, destDir: st
     await mkdir(join(destDir, "public"));
     await copy(join(sourceDir, "public"), join(destDir, "public"));
   }
+
+  // Add the `sharp` library if `/app` folder exists (i.e. Next.js 13+)
+  // or usesNextImage in `export-marker.json` is set to true.
+  // As of (10/2021) the new Next.js 13 route is in beta, and usesNextImage is always being set to false
+  // if the image component is used in pages coming from the new `/app` routes.
+  if (
+    !(await hasUnoptimizedImage(sourceDir, distDir)) &&
+    (usesAppDirRouter(sourceDir) || (await usesNextImage(sourceDir, distDir)))
+  ) {
+    packageJson.dependencies["sharp"] = "latest";
+  }
+
   await mkdirp(join(destDir, distDir));
   await copy(join(sourceDir, distDir), join(destDir, distDir));
   return { packageJson, frameworksEntry: "next.js" };
@@ -300,14 +388,28 @@ export async function ɵcodegenFunctionsDirectory(sourceDir: string, destDir: st
 /**
  * Create a dev server.
  */
-export async function getDevModeHandle(dir: string) {
+export async function getDevModeHandle(dir: string, hostingEmulatorInfo?: EmulatorInfo) {
+  // throw error when using Next.js middleware with firebase serve
+  if (!hostingEmulatorInfo) {
+    if (await isUsingMiddleware(dir, true)) {
+      throw new FirebaseError(
+        `${clc.bold("firebase serve")} does not support Next.js Middleware. Please use ${clc.bold(
+          "firebase emulators:start"
+        )} instead.`
+      );
+    }
+  }
+
   const { default: next } = relativeRequire(dir, "next");
   const nextApp = next({
     dev: true,
     dir,
+    hostname: hostingEmulatorInfo?.host,
+    port: hostingEmulatorInfo?.port,
   });
   const handler = nextApp.getRequestHandler();
   await nextApp.prepare();
+
   return (req: IncomingMessage, res: ServerResponse, next: () => void) => {
     const parsedUrl = parse(req.url!, true);
     const proxy = createServerResponseProxy(req, res, next);